# Conflicts:
#	src/main/java/com/ii/testautomation/service/SchedulingService.java
#	src/main/java/com/ii/testautomation/service/impl/SchedulingServiceImpl.java

+++ src/main/java/com/ii/testautomation/controllers/ExecutionHistoryController.java
@@ -4,15 +4,22 @@ import com.ii.testautomation.enums.RequestStatus;
 import com.ii.testautomation.response.common.BaseResponse;
 import com.ii.testautomation.response.common.ContentResponse;
 import com.ii.testautomation.service.ExecutionHistoryService;
+import com.ii.testautomation.service.ProjectService;
 import com.ii.testautomation.service.TestGroupingService;
 import com.ii.testautomation.utils.Constants;
 import com.ii.testautomation.utils.EndpointURI;
 import com.ii.testautomation.utils.StatusCodeBundle;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
+import java.sql.Date;
+import java.sql.Timestamp;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.LocalDateTime;
 
 @RestController
 @CrossOrigin
@@ -21,6 +28,8 @@ public class ExecutionHistoryController {
     private ExecutionHistoryService executionHistoryService;
     @Autowired
     private TestGroupingService testGroupingService;
+    @Autowired
+    private ProjectService projectService;
     @Autowired
     private StatusCodeBundle statusCodeBundle;
 
@@ -54,14 +63,44 @@ public class ExecutionHistoryController {
         return ResponseEntity.ok(response);
     }
 
+    @GetMapping(EndpointURI.EXECUTION_HISTORY_DATE_FILTER)
+    public ResponseEntity<Object> executionHistoryDateFilter(@PathVariable Long id,@RequestParam(value = "startDate",required = false) String startDate,
+                                                             @RequestParam(value = "endDate",required = false) String endDate) throws ParseException {
+
+         if (!testGroupingService.existsByTestGroupingId(id))
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(), statusCodeBundle.getTestGroupingNotExistCode(), statusCodeBundle.getTestGroupingNotExistsMessage()));
+        if (!executionHistoryService.existByTestGropingId(id))
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(), statusCodeBundle.getFailureCode(), statusCodeBundle.getTestGroupingNotMappedMessage()));
+
+        Timestamp startingDate;
+        Timestamp endingDate;
+
+        if (startDate.isEmpty() || startDate.isBlank())
+        {
+                startingDate = Timestamp.valueOf(LocalDateTime.now().withDayOfMonth(1));
+                endingDate = Timestamp.valueOf(LocalDateTime.now());
+        }
+        else {
+            startingDate = new Timestamp(Date.valueOf(startDate).getTime());
+            endingDate = new Timestamp(Date.valueOf(endDate).getTime());
+        }
+
+        return ResponseEntity.ok(new ContentResponse<>(Constants.EXECUTION_HISTORY,executionHistoryService.executionHistoryDateFilter(id,startingDate,endingDate),RequestStatus.SUCCESS.getStatus(), statusCodeBundle.getCommonSuccessCode(),statusCodeBundle.getViewExecutionHistoryMessage()));
+
+    }
+
     @DeleteMapping(value = EndpointURI.EXECUTION_HISTORY_PROJECT_ID)
-    public ResponseEntity<Object>deleteExecutionHistoryById(@PathVariable Long id,@PathVariable Long projectId)
-    {
+    public ResponseEntity<Object>deleteExecutionHistoryById(@PathVariable Long id,@PathVariable Long projectId) {
         if(!executionHistoryService.existByExecutionHistoryId(id))
         {
-            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(), statusCodeBundle.getTestGroupingNotExistCode(),statusCodeBundle.getTestGroupingNotExistsMessage()));
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(), statusCodeBundle.getExecutionHistoryNotExistsCode(),statusCodeBundle.getExecutionHistoryNotFound()));
+        }
+        if(!projectService.existByProjectId(projectId))
+        {
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(), statusCodeBundle.getProjectNotExistCode(),statusCodeBundle.getProjectNotExistsMessage()));
         }
         executionHistoryService.deleteExecutionHistory(id,projectId);
         return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(), statusCodeBundle.getCommonSuccessCode(),statusCodeBundle.getExecutionHistoryDeleteSuccessMessage()));
     }
+
 }
